+++ src/Service/CronTaskHandler.php
@@ -19,7 +19,6 @@ class CronTaskHandler extends ScheduledTaskHandler
     private EntityRepositoryInterface $salesChannelRepository;
     private SystemConfigService $systemConfigService;
     private Logger $logger;
-    private EntityRepositoryInterface $orderTransactionRepository;
     private EntityRepositoryInterface $orderRepository;
     private EntityRepositoryInterface $customerRepository;
     private OrderTransactionStateHandler $transactionStateHandler;
@@ -30,7 +29,6 @@ class CronTaskHandler extends ScheduledTaskHandler
         EntityRepositoryInterface    $salesChannelRepository,
         SystemConfigService          $systemConfigService,
         Logger                       $logger,
-        EntityRepositoryInterface    $orderTransactionRepository,
         EntityRepositoryInterface    $orderRepository,
         EntityRepositoryInterface    $customerRepository,
         OrderTransactionStateHandler $transactionStateHandler,
@@ -40,7 +38,6 @@ class CronTaskHandler extends ScheduledTaskHandler
         $this->salesChannelRepository = $salesChannelRepository;
         $this->systemConfigService = $systemConfigService;
         $this->logger = $logger;
-        $this->orderTransactionRepository = $orderTransactionRepository;
         $this->orderRepository = $orderRepository;
         $this->customerRepository = $customerRepository;
         $this->transactionStateHandler = $transactionStateHandler;
@@ -98,7 +95,7 @@ class CronTaskHandler extends ScheduledTaskHandler
                     $qb->expr()->like('ot.custom_fields', "'%payment_transaction_status\": \"56%'")
                 )
             )
-            ->orderBy('ot.updated_at','DESC')
+            ->orderBy('ot.updated_at', 'DESC')
             ->setParameter('salesChannelId', $salesChannelId);
 
         if ($daysInterval > 0) {
@@ -133,24 +130,24 @@ class CronTaskHandler extends ScheduledTaskHandler
         }
         $hostedCheckoutId = $customFields[Form::CUSTOM_FIELD_WORLDLINE_PAYMENT_HOSTED_CHECKOUT_ID];
 
-        $orderTransaction = PaymentHandler::getOrderTransaction(
+        $order = PaymentHandler::getOrder(
             Context::createDefaultContext(),
-            $this->orderTransactionRepository,
+            $this->orderRepository,
             $hostedCheckoutId
         );
 
         $paymentHandler = new PaymentHandler(
             $this->systemConfigService,
             $this->logger,
-            $orderTransaction,
+            $order,
             $this->translator,
             $this->orderRepository,
-            $this->orderTransactionRepository,
             $this->customerRepository,
             Context::createDefaultContext(),
             $this->transactionStateHandler
         );
 
-        $paymentHandler->capturePayment($hostedCheckoutId);
+        $amount = (int)round($order->getAmountTotal() * 100);
+        $paymentHandler->capturePayment($hostedCheckoutId, $amount, []);
     }
 }