+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.5.0
+- Order item statuses
+
+ # 1.4.0
+- Update statuses of changed log data
+ 
+# 1.3.0
+- Partial capture, cancel, refund
+- Transaction log
+
 # 1.2.0
 - Klarna and Oney payment methods